Implement end to end D2D1 pixel shader effect tests
